requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
pydantic==2.5.0
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
xgboost==2.0.3
groq  